Use nonbreaking space between status and package
In the line that displays the status and the packagename, use a &nbsp;
instead of a regular space to not make it wrap in between. It would gain
only two characters in length but on the other hand spawn a second line
in case the display is narrower than to view it all.